pyproject.toml
[tool.poetry]
name = "reasoning-forest"
version = "0.1.0"
description = "Interpretable tree/forest-style decision-making systems powered by LLMs"
authors = ["Vela Research <researcher@vela.partners>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "reasoning_forest" }]
keywords = ["decision-trees", "explainable-ai", "llm", "machine-learning", "research"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[tool.poetry.dependencies]
python = ">=3.13,<3.14"
pydantic = "^2.11.7"
pydantic-settings = "^2.10.1"
google-genai = "^1.33.0"
openai = "^1.106.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-asyncio = "^1.1.0"
ruff = "^0.12.12"
pyright = "^1.1.405"
sphinx = "^8.2.3"
pre-commit = "^4.3.0"

[tool.poetry.group.docs.dependencies]
furo = "^2025.7.19"   # clean theme for Sphinx
myst-parser = "^4.0.1"  # if you want Markdown docs

[build-system]
requires = ["poetry-core>=2.1.3"]
build-backend = "poetry.core.masonry.api"
